+++ app/src/main/java/com/depromeet/baton/presentation/ui/search/SearchDetailFragment.kt
@@ -1,0 +1,51 @@
+package com.depromeet.baton.presentation.ui.search
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.View
+import androidx.recyclerview.widget.GridLayoutManager
+import com.depromeet.baton.R
+import com.depromeet.baton.databinding.FragmentSearchDetailBinding
+import com.depromeet.baton.presentation.base.BaseFragment
+import com.depromeet.baton.presentation.ui.detail.TicketDetailActivity
+import com.depromeet.baton.presentation.ui.home.adapter.TicketItemRvAdapter
+import com.depromeet.baton.presentation.ui.home.view.TicketItem
+import com.depromeet.baton.presentation.util.TicketItemVerticalDecoration
+
+class SearchDetailFragment : BaseFragment<FragmentSearchDetailBinding>(R.layout.fragment_search_detail){
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setTicketItemRvAdapter()
+    }
+
+    private fun setTicketItemRvAdapter() {
+        with(binding) {
+            val ticketItemRvAdapter =
+                TicketItemRvAdapter(TicketItemRvAdapter.SCROLL_TYPE_VERTICAL, requireContext(), ::setTicketItemClickListener)
+            val gridLayoutManager = GridLayoutManager(requireContext(), 2, GridLayoutManager.VERTICAL, false)
+
+            adapter = ticketItemRvAdapter
+            itemDecoration = TicketItemVerticalDecoration()
+            rvHome.layoutManager = gridLayoutManager
+
+            ticketItemRvAdapter.submitList(
+                arrayListOf(
+                    TicketItem(
+                        "휴메이크 휘트니스 석촌점", "헬스", "123,000원", "50일 남음", "광진구 중곡동", "12m", R.drawable.dummy1),
+                    TicketItem("테리온 휘트니스 당산점", "기타", "100,000원", "30일 남음", "영등포구 양평동", "12m", R.drawable.dummy2),
+                    TicketItem("진휘트니스 양평점", "헬스", "3,000원", "60일 남음", "광진구 중곡동", "12m", R.drawable.dummy3),
+                    TicketItem("휴메이크 휘트니스 석촌점", "필라테스", "223,000원", "4일 남음", "광진구 중곡동", "12m", R.drawable.dummy4),
+                    TicketItem("바톤휘트니스 대왕점", "헬스", "19,000원", "5일 남음", "광진구 중곡동", "12m", R.drawable.dummy5),
+                    TicketItem("휴메이크 휘트니스 석촌점", "필라테스", "223,000원", "4일 남음", "광진구 중곡동", "12m", R.drawable.dummy7),
+                )
+            )
+        }
+    }
+
+    private fun setTicketItemClickListener(ticketItem: TicketItem) {
+        startActivity(Intent(requireContext(), TicketDetailActivity::class.java).apply {
+            //TODO 게시글 id넘기기
+        })
+    }
+}
