+++ app/src/main/java/com/depromeet/baton/presentation/ui/search/SearchFragment.kt
@@ -3,17 +3,24 @@ package com.depromeet.baton.presentation.ui.search
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.repeatOnLifecycle
 import com.depromeet.baton.R
 import com.depromeet.baton.databinding.FragmentSearchBinding
 import com.depromeet.baton.presentation.base.BaseFragment
-import com.depromeet.bds.component.BdsSearchBar
+import com.depromeet.baton.presentation.util.viewLifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.launch
 import timber.log.Timber
 
 @AndroidEntryPoint
 class SearchFragment : BaseFragment<FragmentSearchBinding>(R.layout.fragment_search) {
 
     private val viewModel: SearchViewModel by activityViewModels()
+
+    private val recentFragment = RecentSearchFragment()
+    private val detailFragment = SearchDetailFragment()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -23,9 +30,24 @@ class SearchFragment : BaseFragment<FragmentSearchBinding>(R.layout.fragment_sea
             viewModel.searchKeyword(it.text.toString())
         }
 
-        childFragmentManager
-            .beginTransaction()
-            .replace(R.id.container_fragment, RecentSearchFragment(), null)
-            .commit()
+        viewLifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.searchKeyword
+                    .map { it.isBlank() }
+                    .collect { showRecent ->
+                        if (showRecent) {
+                            childFragmentManager
+                                .beginTransaction()
+                                .replace(R.id.container_fragment, recentFragment, null)
+                                .commit()
+                        } else {
+                            childFragmentManager
+                                .beginTransaction()
+                                .replace(R.id.container_fragment, detailFragment, null)
+                                .commit()
+                        }
+                    }
+            }
+        }
     }
 }
