+++ app/src/main/java/com/depromeet/baton/presentation/util/CoroutineExt.kt
@@ -1,0 +1,8 @@
+package com.depromeet.baton.presentation.util
+
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.LifecycleCoroutineScope
+import androidx.lifecycle.lifecycleScope
+
+val Fragment.viewLifecycleScope: LifecycleCoroutineScope
+    get() = viewLifecycleOwner.lifecycleScope
