+++ app/src/main/java/com/depromeet/baton/util/SimpleDiffUtil.kt
@@ -1,0 +1,15 @@
+package com.depromeet.baton.util
+
+import android.annotation.SuppressLint
+import androidx.recyclerview.widget.DiffUtil
+
+class SimpleDiffUtil<T : Any> : DiffUtil.ItemCallback<T>() {
+    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem == newItem
+    }
+
+    @SuppressLint("DiffUtilEquals")
+    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
+        return oldItem == newItem
+    }
+}
