+++ app/src/main/res/layout/fragment_search_detail.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+        <variable
+            name="itemDecoration"
+            type="com.depromeet.baton.presentation.util.TicketItemVerticalDecoration" />
+
+        <variable
+            name="adapter"
+            type="com.depromeet.baton.presentation.ui.home.adapter.TicketItemRvAdapter" />
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/fcv_home"
+            android:name="com.depromeet.baton.presentation.ui.filter.view.FilterChipFragment"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="@dimen/dimen12"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rv_home"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:layout_marginHorizontal="@dimen/dimen16"
+            android:layout_marginTop="@dimen/dimen12"
+            android:nestedScrollingEnabled="false"
+            app:adapter="@{adapter}"
+            app:itemDecoration="@{itemDecoration}"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/fcv_home"
+            app:spanCount="2"
+            tools:itemCount="5"
+            tools:listitem="@layout/item_ticket" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
